apps/bot/requirements.txt
python-telegram-bot==21.3
flask==3.0.3
gunicorn==22.0.0
supabase==2.5.0
google-generativeai==0.7.2
python-dotenv==1.0.1
